update & zoeken

+++ Model/QueryBuilder/SponsorQueryBuilder.php
@@ -9,14 +9,29 @@
 class SponsorQueryBuilder
 {
 
-    public function getAllSponsors()
+    public function getAllSponsors($keyword = null)
     {
-        return Database::query("SELECT * FROM `sponsor`");
+        $sql = "SELECT * FROM `sponsor` as s";
+
+        if ($keyword != null) {
+            $sql .= " WHERE s.Name LIKE ? 
+                  OR s.WebsiteLink LIKE ?
+                  OR s.user_Email LIKE ?
+                  OR s.Description LIKE ?";
+            $parameters = array($keyword, $keyword, $keyword, $keyword);
+        }
+
+        if (isset($parameters)) {
+            return Database::query_safe($sql, $parameters);
+        } else {
+            return Database::query($sql);
+        }
     }
 
 
     public function addSponsor(Sponsor $sponsor)
     {
+        if ($sponsor == null) return;
         $sql = "INSERT INTO `aladdin_db2`.`sponsor`";
         $sql .= " (`Name`, `Image`, `Description`, `WebsiteLink`, `user_Email`)";
         $sql .= " VALUES (?,?,?,?,?)";
@@ -24,9 +39,22 @@ class SponsorQueryBuilder
         Database::query_safe($sql, $parameters);
     }
 
-    public function deleteSponsor(Sponsor $sponsor){
+    public function updateSponsor(Sponsor $sponsor)
+    {
+        print_r($sponsor);
+        exit();
+        if ($sponsor == null) return;
+        $sql = "UPDATE `sponsor` SET `Name` = ?,`Image`=?,`Description`=?,`WebsiteLink`=?,`user_Email` = ? WHERE Id = ?;";
+        $parameters = array($sponsor->name, $sponsor->image, $sponsor->description, $sponsor->url, $sponsor->userMail, $sponsor->id);
+        Database::query_safe($sql, $parameters);
+
+    }
+
+    public function deleteSponsor(Sponsor $sponsor)
+    {
+        if ($sponsor == null) return;
         $sql = "DELETE FROM `aladdin_db2`.`sponsor` WHERE `Id` = ?";
         Database::query_safe($sql, array($sponsor->id));
     }
-    
+
 }
